move the version parsing to a separate file
