+++ gradle-plugin/plugin/src/main/java/com/yelp/codegen/plugin/GenerateTaskConfiguration.kt
@@ -1,7 +1,5 @@
 package com.yelp.codegen.plugin
 
-import io.swagger.parser.SwaggerParser
-import java.io.File
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.Project
@@ -23,22 +21,4 @@ abstract class GenerateTaskConfiguration @Inject constructor(project: Project) {
     val features: FeatureConfiguration = FeatureConfiguration(objects)
 
     fun features(action: Action<FeatureConfiguration>) = action.execute(features)
-
-    companion object {
-        private fun readVersionFromSpecfile(specFile: File): String {
-            val swaggerSpec = SwaggerParser().readWithInfo(specFile.absolutePath, listOf(), false).swagger
-
-            return when (val version = swaggerSpec.info.version) {
-                is String -> {
-                    println("Successfully read version from Swagger Spec file: $version")
-                    version
-                }
-                else -> {
-                    val defaultVersion = "0.0.0"
-                    println("Issue in reading version from Swagger Spec file. Falling back to $defaultVersion")
-                    defaultVersion
-                }
-            }
-        }
-    }
 }
